moved state files into new state folder
